requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Sheets API
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# Data processing - Python 3.13 compatible versions
pandas>=2.2.0
numpy>=1.26.4

# Machine Learning
scikit-learn>=1.3.2
joblib==1.3.2

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment management
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Date handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
